Rename PingPacket to GameStateRequest
